separate some constant names to common place
