+++ sdks/node/src/integrations/index.ts
@@ -5,5 +5,5 @@ export {
   instrumentOpenAIAgents,
 } from './openai.agent';
 export {patchRealtimeSession} from './openai.realtime.agent';
-export {createOtelExtension, PiCodingAgentOtelAdapter} from './pi.coding.agent';
-export type {OtelExtensionOptions} from './pi.coding.agent';
+export {createOtelExtension, PiCodingAgentOtelAdapter} from './piCodingAgent';
+export type {OtelExtensionOptions} from './piCodingAgent';
