+++ sdks/node/src/integrations/piCodingAgent.ts
@@ -29,6 +29,8 @@ import {
   trace,
 } from '@opentelemetry/api';
 
+import {GEN_AI_ATTR, OTEL_ATTR} from './common/genai';
+
 import type {
   PiAgentMessage,
   PiAssistantMessage,
@@ -37,35 +39,25 @@ import type {
   PiExtensionDefinition,
   PiExtensionEvent,
   PiModel,
-} from './pi.coding.agent.types';
+} from './piCodingAgent.types';
 
 // ---------------------------------------------------------------------------
-// GenAI semantic convention attribute keys
+// Options
+// ---------------------------------------------------------------------------
+
+/** Options for the pi coding agent OTEL extension. */
+export interface OtelExtensionOptions {
+  tracer?: Tracer;
+  captureContent?: boolean;
+}
+
+// ---------------------------------------------------------------------------
+// Attribute keys — GenAI semconv keys from common, pi-specific keys below
 // ---------------------------------------------------------------------------
 
 const ATTR = {
-  // GenAI attributes
-  GEN_AI_PROVIDER_NAME: 'gen_ai.provider.name',
-  GEN_AI_OPERATION_NAME: 'gen_ai.operation.name',
-  GEN_AI_AGENT_NAME: 'gen_ai.agent.name',
-  GEN_AI_REQUEST_MODEL: 'gen_ai.request.model',
-  GEN_AI_RESPONSE_MODEL: 'gen_ai.response.model',
-  GEN_AI_RESPONSE_FINISH_REASONS: 'gen_ai.response.finish_reasons',
-  // Conversation / session correlation
-  GEN_AI_CONVERSATION_ID: 'gen_ai.conversation.id',
-  // Token usage
-  GEN_AI_USAGE_INPUT_TOKENS: 'gen_ai.usage.input_tokens',
-  GEN_AI_USAGE_OUTPUT_TOKENS: 'gen_ai.usage.output_tokens',
-  GEN_AI_USAGE_TOTAL_TOKENS: 'gen_ai.usage.total_tokens',
-  GEN_AI_USAGE_CACHE_READ_INPUT_TOKENS: 'gen_ai.usage.cache_read.input_tokens',
-  GEN_AI_USAGE_CACHE_CREATION_INPUT_TOKENS:
-    'gen_ai.usage.cache_creation.input_tokens',
-  // Tool
-  GEN_AI_TOOL_NAME: 'gen_ai.tool.name',
-  GEN_AI_TOOL_CALL_ID: 'gen_ai.tool.call.id',
-  GEN_AI_OUTPUT_TYPE: 'gen_ai.output.type',
-  // Error
-  ERROR_TYPE: 'error.type',
+  ...GEN_AI_ATTR,
+  ...OTEL_ATTR,
   // Pi-specific (not in GenAI spec)
   PI_SESSION_CWD: 'pi.session.cwd',
   PI_USAGE_COST_USD: 'pi.usage.cost_usd',
@@ -105,25 +97,6 @@ function resolveGenAiProviderName(provider: string): string {
 /** Narrows a PiAgentMessage to PiAssistantMessage, or returns undefined. */
 function asAssistant(msg: PiAgentMessage): PiAssistantMessage | undefined {
   return msg.role === 'assistant' ? (msg as PiAssistantMessage) : undefined;
-}
-
-// ---------------------------------------------------------------------------
-// Options
-// ---------------------------------------------------------------------------
-
-export interface OtelExtensionOptions {
-  /**
-   * OTEL Tracer to use for creating spans.
-   * Defaults to trace.getTracer('pi-coding-agent-weave-ext') which reads
-   * from whatever TracerProvider the user has configured globally.
-   */
-  tracer?: Tracer;
-  /**
-   * Whether to attach message text as gen_ai.event.content on span events.
-   * When false, events are still emitted (for timing/flow) but without content.
-   * Default: true.
-   */
-  captureContent?: boolean;
 }
 
 // ---------------------------------------------------------------------------
@@ -285,4 +258,4 @@ export type {
   PiExtensionDefinition,
   PiExtensionContext,
   PiModel,
-} from './pi.coding.agent.types';
+} from './piCodingAgent.types';
